Second round of windows compatibility fixes.
Added GLFW OpenGL context, removed ConditionVariable.

+++ CoreRender/src/core/FrameLimiter.cpp
@@ -52,7 +52,7 @@ namespace core
 		if (currenttime < lasttime)
 		{
 			// TODO: Statistics
-			Time::sleep(lasttime - currenttime);
+			Time::sleep((unsigned int)(lasttime - currenttime));
 		}
 		else
 		{
